back/requirements.txt
django
djangorestframework
ipython
django-extensions
Werkzeug
requests
django-filter
